summarization/sum_flask/requirements.txt
Flask
pandas
PyMuPDF
torch==1.8.1
sentence-transformers==1.1.0
spacy==2.3.1
